back-end/requirements.txt
Flask
pymongo
flask-cors
azure-storage
celery
redis
minio
pyjwt
requests
neo4j
